+++ tests/examples/setup_0.0.0.py
@@ -3,6 +3,7 @@ from setuptools import setup
 setup(
     name='package',
     version='0.0.0',
+    no_version='1.1.1',
     description='Description',
     author='Javier Santacruz',
     author_email='jsl@taric.es'