+++ tests/test_cli.py
@@ -6,7 +6,7 @@ import tempfile
 import contextlib
 
 from tests import CommandTest, here
-from hamcrest import assert_that, contains_string
+from hamcrest import assert_that, contains_string, matches_regexp
 
 EXAMPLE = here('examples', 'setup_0.0.0.py')
 
@@ -20,7 +20,16 @@ def version_file(version='0.0.0'):
         yield tmp.name
 
 
+class TestRootCommand(CommandTest):
+    def test_it_shows_version(self):
+        self.run(['--version'])
+
+        self.assert_result(output=matches_regexp(r'version \d+\.\d+\.\d+.*'))
+
+
 class TestView(CommandTest):
+    defargs = ['--path', EXAMPLE]
+
     def test_it_should_update_major(self):
         self.run(['view'] + ['--major'])
 
@@ -31,8 +40,8 @@ class TestView(CommandTest):
 
         self.assert_result(output=contains_string('From 0.0.0 to 0.1.0'))
 
-    def test_it_should_update_patch(self):
-        self.run(['view'] + ['--patch'])
+    def test_it_should_update_revision(self):
+        self.run(['view'] + ['--revision'])
 
         self.assert_result(output=contains_string('From 0.0.0 to 0.0.1'))
 
@@ -46,10 +55,30 @@ class TestView(CommandTest):
 
         self.assert_result(output=contains_string('From 0.0.0 to 0.0.0.post1'))
 
+    def test_it_should_update_major_and_revision(self):
+        self.run(['view'] + ['--major', '--revision'])
+
+        self.assert_result(output=contains_string('From 0.0.0 to 1.0.1'))
+
+    def test_it_should_update_minor_and_revision(self):
+        self.run(['view'] + ['--minor', '--revision'])
+
+        self.assert_result(output=contains_string('From 0.0.0 to 0.1.1'))
+
+    def test_it_should_update_major_and_minor_and_revision(self):
+        self.run(['view'] + ['--major', '--minor', '--revision'])
+
+        self.assert_result(output=contains_string('From 0.0.0 to 1.1.1'))
+
     def test_it_should_update_all_at_the_same_time(self):
-        self.run(['view'] + ['--major', '--minor', '--patch', '--dev', '--post'])
+        self.run(['view'] + ['--major', '--minor', '--revision', '--dev', '--post'])
 
         self.assert_result(output=contains_string('From 0.0.0 to 1.1.1.post1.dev1'))
+
+    def test_it_should_update_given_var(self):
+        self.run(['view'] + ['--major', '--minor', '--revision', '--var', 'no_version'])
+
+        self.assert_result(output=contains_string('From 1.1.1 to 2.1.1'))
 
     def test_it_should_remove_dev_when_final(self):
         with version_file('0.0.0.dev1') as path:
@@ -80,7 +109,7 @@ class TestUp(CommandTest):
             example_path = os.path.join(path, 'setup.py')
             shutil.copyfile(EXAMPLE, example_path)
 
-            self.run(['up', '--major', '--minor', '--patch', '--dev', '--post'])
+            self.run(['up', '--major', '--minor', '--revision', '--dev', '--post'])
 
             self.assert_result(output=contains_string('From {} to {}'.format(initial, expected)))
             self.assert_result(output=contains_string('writing "{}"'.format(example_path)))