+++ HISTORY.rst
@@ -1,7 +1,14 @@
-0.3.0 (2015-10-08)
+0.4.0 (2015-10-08)
 ------------------
 
 * Adds --final option which removes .dev or .post suffix
+
+0.3.0 (2015-09-22)
+------------------
+
+* Changes --patch option for --revision
+* Adds --version option
+* Fixes problem with version substitution
 
 0.2.1 (2015-09-12)
 ------------------